requirements.txt
Django==1.8
django-tagging==0.3.4
